Finishing create room handler

+++ src/models/room.rs
@@ -8,7 +8,7 @@ use crate::schemas::room_schema::RoomStatus;
 pub struct Room {
     pub id: Uuid,
     pub kost_id: Uuid,
-    pub room_number: u8,
+    pub room_number: u32,
     pub room_vacancy: RoomStatus,
     pub created_at: Option<DateTime<Utc>>,
     pub updated_at: Option<DateTime<Utc>>
